+++ src/main/java/com/danteck/studentmanagementsysem/service/impl/StudentServiceImpl.java
@@ -22,22 +22,27 @@ public class StudentServiceImpl implements StudentService {
     }
 
     @Override
-    public Student saveStudent(Student student) {
-        return studentRepository.save(student);
+    public void saveStudent(Student student) {
+        studentRepository.save(student);
     }
 
     @Override
     public Student getStudentById(Long id) {
-        return studentRepository.findById(id).get();
+        return studentRepository.findById(id).orElse(null);
     }
 
     @Override
-    public Student updateStudent(Student student) {
-        return studentRepository.save(student);
+    public void updateStudent(Student student) {
+        studentRepository.save(student);
     }
 
     @Override
     public void deleteStudentById(Long id) {
-        studentRepository.deleteById(id);
+        if (getStudentById(id) == null)
+        {
+
+        } else {
+            studentRepository.deleteById(id);
+        }
     }
 }
